Fix validation of booleans for required fields
If the minimum cardinality of boolean field was 1 (it was required) then
the validation was failing when the value was false. This was caused by
an existence check in the code evaluating to falsy in that case, and can
be avoided by explicitly checking that the value is not undefined.